Rendering and collision detection work
Moving from PositionComponent to AreaComponent being ubiquitous for entities,
this allows the rest of the engine to be agnostic about an entity's size as
long as it's specified in the area.

Also begun implementing rendering the appropriate thing where an entity
used to be, and calculating dirty rectangles for rerendering.