De-factoring input events
I'm currently abstracting input events as a `RawEvent` class with the
intention of allowing the game engine to process events not directly
tied to a local UI (i.e., over the web). However, this incurs a
translation layer that will usually reduce to a
java.awt.event.InputEvent anyway, so let's just assume `InputEvent`s
as our native currency and if more abstract input events need to be
shipped over the web we can write a translation layer to convert those to `InputEvent`s.

Also added dock image for OSX.