Deferred scene (re-)initialization
Upon game startup, we need to give PossumSceneBuilder all of the scenes
we intend to use, but they may require expensive initialization and we
don't want to front-load all of it at once before the first scene is
even rendered. This commit adds a SceneInitializer interface which
will be called before a scene's first frame is run, giving it a chance
to do whatever setup it needs at that time.