Loading screen
For scenes which take a long time to load (generating caves via cellular
automata, etc.), it would be helpful to display an intermediate
"loading" scene while we perform the expensive initialization in the
background.

This is currently implemented by having the loading scene's init
handler fire off the next scene's initialization on another thread,
and registers a handler with the event broker to transition to the
next scene as once it receives the SceneInitializedEvent indicating it's
ready.