Threading bug
Depending on the (nondeterministic) order of operations between
'Thread.start' and 'nextScene = nextScene.loadingScene', there was no
guarantee that the actually be loading the correct scene on the new
thread, in which case we'd essentially lock up there.

Finalizing the nextScene at the point it's correct so we don't end up
with the value being lexically scoped.