Don't add panel contents to quadtree
When adding an AreaComponent to entities belonging to a panel, they
were then being added to the quadtree. Since we don't do collision
detection on panels, we now check to see if an entity is a direct
child of a panel before adding it to the quadtree.

Also eliminated explicit entity initialization as it's no longer needed.