Pause System decoupled from Scene
Scenes no longer explictly hold state for the pause system. In line with
the direction we've gone with the other systems, it is as orthogonal and
separate from the Scene's state as possible.

The SwingKeyboardInputAdapter still retains some coupling to the pause
system in that it needs to act on the Event Dispatch Thread.