Rendering quadtree outlines
When enabling debug mode, paints the boundaries of each visibile
quadtree node.  On receipt of a pre-render event, we know to
repaint any boundaries that may have been obscured.

NOTE: this paints every node from the root down to the lowest level
the entity is contained in, but we're not too worried about efficiency here.

ALSO NOTE: This isn't perfect yet, when an entity moves away from a boundary
edge, the boundary may be painted and then overwritten shortly after by
whatever was behind the entity (e.g., a floor tile). Good enough for now.