Fixed room-filling bugs
Adjacent rooms can share walls, but those cells weren't being revisited
on subsequent flood fills. Resetting between fills.

We'd also occasionally miss a wall at the end of a row, so those are now
examined as well.

It would be great to be able to test these, but doing that easily and
in a decipherable manner would probably require a way to specify a
non-random cave layout in ASCII. As we don't yet have a loader for that,
this will have to do for now.